src/Sportradar.LiveOddsService.Business/MatchService.cs
using System;
using System.Threading.Tasks;
using Sportradar.LiveOddsService.Data;
using Sportradar.LiveOddsService.Domain.Models;
using Sportradar.LiveOddsService.Domain.Services;

namespace Sportradar.LiveOddsService.Business {
    public class MatchService: IMatchService {
        private IMatchRepository _matchRepository;

        public MatchService(IMatchRepository matchRepository) => this._matchRepository = matchRepository;

        public Task<Match> StartAsync(string homeTeam, string awayTeam) => throw new NotImplementedException();
    }
}
